HTBHF-2520 move error formatting to be accesible by middleware
