+++ src/web/routes/application/errors/index.js
@@ -1,0 +1,5 @@
+const { wrapError } = require('./errors')
+
+module.exports = {
+  wrapError
+}
