+++ src/web/routes/application/steps/common/formatters/index.js
@@ -1,4 +1,3 @@
 module.exports = {
-  ...require('./dates'),
-  ...require('./errors')
+  ...require('./dates')
 }
