+++ src/web/routes/application/steps/confirm/confirm.js
@@ -3,7 +3,7 @@ const { path, isNil } = require('ramda')
 const { handleRequestForPath } = require('../../flow-control')
 const { CONFIRM_URL } = require('../../paths')
 const { ELIGIBLE } = require('../common/constants')
-const { wrapError } = require('../common/formatters')
+const { wrapError } = require('../../errors')
 
 const toPounds = pence => (parseInt(pence, 10) / 100).toFixed(2)
